+++ src/obj/vendor.h
@@ -9,8 +9,10 @@
 #define LUMINARI_OBJ_VENDOR_H
 
 struct char_data;
+struct spec_event_context;
 
 int bank(struct char_data *ch, void *me, int cmd, const char *argument);
+int bank_typed(struct spec_event_context *context);
 int bought_pet(struct char_data *ch, void *me, int cmd, const char *argument);
 int buyarmor(struct char_data *ch, void *me, int cmd, const char *argument);
 int buyweapons(struct char_data *ch, void *me, int cmd, const char *argument);